+++ package.json
@@ -2,14 +2,23 @@
   "name": "NodeAPI",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "test": "jasmine-node tests"
+    "test": "jasmine-node tests",
+    "start": "node app.js"
   },
-  "keywords": [],
+  "keywords": [
+    "heroku"
+  ],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.16.4"
+    "express": "^4.16.4",
+    "joi": "^14.3.1",
+    "l": "^0.6.0",
+    "mongodb": "^3.2.2",
+  },
+  "engines": {
+    "node": "10.15.3"
   }
 }
